Exception handling

+++ src/main/java/com/lutheone/dscommerce/services/ProductService.java
@@ -3,9 +3,14 @@ package com.lutheone.dscommerce.services;
 import com.lutheone.dscommerce.dto.ProductDTO;
 import com.lutheone.dscommerce.entities.Product;
 import com.lutheone.dscommerce.repositories.ProductRepository;
+import com.lutheone.dscommerce.services.exceptions.DataBaseExeception;
+import com.lutheone.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 
@@ -22,7 +27,7 @@ public class ProductService {
     public ProductDTO findById(Long id) {
         return repository.findById(id)
                 .map(ProductDTO::new)
-                .orElseThrow(() -> new RuntimeException("Produto não encontrado"));
+                .orElseThrow(() -> new ResourceNotFoundException("Entity not found"));
 
     }
 
@@ -44,16 +49,30 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product entity = repository.findById(id).orElse(null);
-        copyDtoToEntity(productDTO, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
-        }
 
-        @Transactional
-        public void deleteById(Long id) {
-        repository.deleteById(id);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(productDTO, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
         }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Id not found " + id);
+        }
+    }
+
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource not found with id: " + id);
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DataBaseExeception("Cannot delete resource with id: " + id + " due to integrity violation");
+        }
+    }
 
     private void copyDtoToEntity(ProductDTO productDTO, Product entity) {
         entity.setName(productDTO.getName());
